PrescriptionCreate completed and DataGenerate almost finalized
PrescriptionCreate is completely functional. prescription _show.html has been adjusted to reflect the way the database handles formatting of addresses.

DataGenerate is functional. Still have to make it so that it only generates *only correct* SSN keys for patients and doctors, as well as change the way they are checked to see if they are unique by adding a HashSet of old SSN values.